Refactors the importer for S3 to inherit from remote csv importer and override some of the behaviors, extracts most of the logic to the parent class including CSV options and data type conversion methods
