+++ spec/fixtures/files/wonderschool_necc_attendance_data_missing_field.csv
@@ -1,9 +1,9 @@
-attendance_id,child_id,school_name,checked_out_at
-5772c9a6-c675-47e0-b8a3-923e076210f5,1234,Test Daycare,
-0522fe23-3daa-465f-afd4-e23c01f6e11e,5678,Fake Daycare,2021-02-24 22:35:29+00
-dc3e6cfe-2dbc-4f16-a8e6-bd202e8600b5,5678,Fake Daycare,2021-03-03 13:06:59+00
-a245d254-be43-4635-9708-c3b8fb6e3dba,1234,Test Daycare,2021-03-04 19:59:28+00
-24a846d4-b373-43f7-8791-dc6f4df05ea6,1234,Test Daycare,2021-03-05 18:23:43+00
-fb373063-4637-4c07-b5b2-d8b9a56d08df,1234,Test Daycare,
-f5e8609b-d39c-4efb-b098-8eb2ccddc8b5,5677,Fake Daycare,2021-03-11 00:27:53+00
-138cd832-4f35-4812-b545-b625d5155f88,5677,Fake Daycare,2021-03-11 23:02:34+00
+child_id,checked_out_at,attendance_id,edited_check_in,edited_check_out,original_checked_in_at,original_checked_out_at
+1234,,5772c9a6-c675-47e0-b8a3-923e076210f5,,,11:23:14,
+5678,2021-02-24 22:35:29+00,0522fe23-3daa-465f-afd4-e23c01f6e11e,,,11:46:43,
+5678,2021-03-03 13:06:59+00,dc3e6cfe-2dbc-4f16-a8e6-bd202e8600b5,,,14:33:58,
+1234,2021-03-04 19:59:28+00,a245d254-be43-4635-9708-c3b8fb6e3dba,,,14:08:31,
+1234,2021-03-05 18:23:43+00,24a846d4-b373-43f7-8791-dc6f4df05ea6,,,22:11:10,22:11:33
+1234,,fb373063-4637-4c07-b5b2-d8b9a56d08df,,,22:12:43,22:13:13
+5677,2021-03-11 00:27:53+00,f5e8609b-d39c-4efb-b098-8eb2ccddc8b5,,,15:14:18,
+5677,2021-03-11 23:02:34+00,138cd832-4f35-4812-b545-b625d5155f88,,,14:33:43,
