+++ spec/services/wonderschool/necc/attendance_csv_importer_spec.rb
@@ -5,15 +5,14 @@ require 'rails_helper'
 module Wonderschool
   module Necc
     RSpec.describe AttendanceCsvImporter do
-      let!(:file_name) { 'file_name.csv' }
-      let!(:source_bucket) { 'source_bucket' }
+      let!(:file_uri) { 'not_a_file.com/file.csv' }
       let!(:archive_bucket) { 'archive_bucket' }
       let!(:akid) { 'akid' }
       let!(:secret) { 'secret' }
       let!(:region) { 'region' }
       let!(:action) { 'action' }
       let!(:stubbed_client) { double('AWS Client') }
-      let!(:stubbed_object) { double('S3 Object') }
+      let!(:stubbed_uri_open) { double('URI Parse and Open') }
 
       let!(:attendance_csv) { File.read(Rails.root.join('spec/fixtures/files/wonderschool_necc_attendance_data.csv')) }
       let!(:invalid_csv) { File.read(Rails.root.join('spec/fixtures/files/invalid_format.csv')) }
@@ -46,27 +45,22 @@ module Wonderschool
         allow(Rails.application.config).to receive(:aws_access_key_id).and_return(akid)
         allow(Rails.application.config).to receive(:aws_region).and_return(region)
         allow(Aws::S3::Client).to receive(:new) { stubbed_client }
-        allow_any_instance_of(described_class).to receive(:source_bucket).and_return(source_bucket)
+        allow(URI).to receive(:parse).with(file_uri) { stubbed_uri_open }
+        allow_any_instance_of(described_class).to receive(:uri).and_return(file_uri)
         allow_any_instance_of(described_class).to receive(:archive_bucket).and_return(archive_bucket)
-        allow(stubbed_client).to receive(:list_objects_v2).with({ bucket: source_bucket }).and_return({ contents: [{ key: file_name }] })
-        allow(stubbed_client).to receive(:get_object).and_return(stubbed_object)
       end
 
       describe '#call' do
         context 'with valid data' do
           before(:each) do
-            allow(stubbed_object).to receive(:body).and_return(attendance_csv)
-            allow(stubbed_client).to receive(:copy_object).and_return({ copy_object_result: {} })
-            allow(stubbed_client).to receive(:delete_object).and_return({})
+            allow(stubbed_uri_open).to receive(:open).and_return(attendance_csv)
+            allow(stubbed_client).to receive(:put_object).and_return({ put_object_result: {} })
           end
 
           it 'creates attendance records for every row in the file, idempotently' do
             expect { described_class.new.call }.to change { Attendance.count }.from(0).to(8)
-            allow(stubbed_client).to receive(:list_objects_v2).with({ bucket: source_bucket }).and_return({ contents: [{ key: file_name }] })
-            allow(stubbed_client).to receive(:get_object).and_return(stubbed_object)
-            allow(stubbed_object).to receive(:body).and_return(attendance_csv)
-            allow(stubbed_client).to receive(:copy_object).and_return({ copy_object_result: {} })
-            allow(stubbed_client).to receive(:delete_object).and_return({})
+            allow(stubbed_uri_open).to receive(:open).and_return(attendance_csv)
+            allow(stubbed_client).to receive(:put_object).and_return({ put_object_result: {} })
             expect { described_class.new.call }.not_to change(Attendance, :count)
           end
 
@@ -82,24 +76,19 @@ module Wonderschool
         end
 
         it "continues processing if the child doesn't exist" do
-          allow(stubbed_object).to receive(:body).and_return(attendance_csv)
+          allow(stubbed_uri_open).to receive(:open).and_return(attendance_csv)
           first_child.destroy!
           described_class.new.call
-          expect(stubbed_client).not_to receive(:copy_object)
-          expect(stubbed_client).not_to receive(:delete_object)
+          expect(stubbed_client).not_to receive(:put_object)
         end
 
         it 'continues processing if the record is invalid or missing a required field' do
-          allow(stubbed_object).to receive(:body).and_return(invalid_csv)
+          allow(stubbed_uri_open).to receive(:open).and_return(invalid_csv)
           described_class.new.call
-          expect(stubbed_client).not_to receive(:copy_object)
-          expect(stubbed_client).not_to receive(:delete_object)
-          allow(stubbed_client).to receive(:list_objects_v2).with({ bucket: source_bucket }).and_return({ contents: [{ key: file_name }] })
-          allow(stubbed_client).to receive(:get_object).and_return(stubbed_object)
-          allow(stubbed_object).to receive(:body).and_return(missing_field_csv)
+          expect(stubbed_client).not_to receive(:put_object)
+          allow(stubbed_uri_open).to receive(:open).and_return(invalid_csv)
           described_class.new.call
-          expect(stubbed_client).not_to receive(:copy_object)
-          expect(stubbed_client).not_to receive(:delete_object)
+          expect(stubbed_client).not_to receive(:put_object)
         end
       end
     end
